[m4] Created m4 macro for gac module checks in configure
This allows dependency checking of dll's that are not distributed
with a .pc file at configure time. Mono.Data:Sqlite is a referenced
dll that needs to be checked, if SQlite backend is enabled. This is
now done via the new macro.

The macro works very much like the PKG_CHECK_MODULES macro. It
uses gacutil to determine if a dll is in the gac or not. It also supports a
custom gacdir via the GACDIR env variable.