[xbuild] Added Substitute task
This task subtitutes values much the same way as autoconf does it. A
substitution can be specified by providing an Item of ItemGroup
"Substitution" with the Include attribute set to:

<pattern>|<value>|<valueType>

where <pattern> is usually something like @prefix@
<value> can be any string - also a path to a text file
<valueType> can be "text" or "path", thus determining if <value> should
be taken directly for substitution or if <value> points to a text file from
which the value for substitution should be taken.

Files that are to be considered in the substitution process must be part
of the "Substitute" ItemGroup