[xbuild] Override BeforeBuild target
The build and AltTarget
-----------------------

This is the main build target of Tasque. Though it primarily serves
as the entry point to compile Tasque, it also provides a way to trigger
other targets if the AltTarget property so indicates.

Explanation: The VS solution file format only allows a limited set of
targets (standard targets). But autotools use a greater variety of targets
(e.g. dist, install, uninstall, ...). Therefore an alternative way to trigger those
targets has been established.

If the target "build" is executed and the "AltTarget" property is empty,
Tasque builds as usual. If however the "AltTarget" property has not empty
and its value represents a valid target in the project file, that target is
executed.

To prohibit interference with the regular build (conducted by
Microsoft.Common.targets proj) during an AltTarget build, the regular build
path is cut short by tweaking the BuildDependsOn property.

A regular build (non-AltTarget build) includes the execution of
	* Substitute
	* GettextCompile
	* Translate
	* BeforeTasqueCompile
before the execution control is handed back to Microsoft.Common.targets.