[xbuild] Add install target
All files that are in the InstallFile ItemGroup are being installed. Those
items also must have the metadata "InstallPath" and "InstallFileName"
set.

A means to add items to "InstallFile" is provided by  a special hook
"SetupInstallFile", which can be overridden by any child project for
that purpose.

Note: The "DestDir" property used in the install destination file path is
required by automake which likes to prepend a tmp path sometimes.