[build] Couple build and install targets
For better integration with MD on Linux, the install target is now
executed autoamtically after build is finished (it's hooked into
AfterBuild). This behavior is controlled by the property "AutoInstall",
which defaults to true. The same goes for the clean and uninstall
targets.

Because this behavior is not expected when doing "configure && \
make && make install" "AutoInstall" is set to false in the Makefile.

To enable debugging/running Tasque from within MD a custom
command has been set up for execution on the tasque project.