[xbuild] Add PackageName and TopBuildDir to each proj
This is necessary to build a proj file separatly. I introduced
this at first, because I thought it would help in building stuff
from within MD, since MD (when the option "Use xbuild" is
set) seemingly invokes xbuild on each proj in the sln instead
of just the sln file.

As it urns out MD does something very wierd. And at this
point I have no idea, why it wouldn't work. I've run it in
diagnostics mode from within MD and it doesn't show any
signs of running the initial target of X.Common.targets.
Though it loads the file correctly and executes other targets
in that file. No clue.