Remove preferences Gtk# dep from model: IBackend.GetPreferencesWidget ()
The method IBackend.GetPreferencesWidget () has been substituted by the read-only property IBackend.Preferences, which is of type IBackendPreferences. This interface has been created to provide an abstract handle for the class PreferencesDialog.cs in the application. A concrete preferences widget is being provided along with the backend implementation: DummyPreferences.cs.

Conflicts:

	src/libtasque/IBackend.cs
	src/libtasque/libtasque.csproj
	src/tasque/tasque.csproj