Task columns are now Mono.Addins
TaskTreeView has been renamed to TaskView and no
longer inherits from TreeView. It has been reduced to
loading and managing the interactions of the various
task columns.

Each task column is now an extension of ITaskColumn,
an interface that provides extensibility via the
Mono.Addins framework.

Two Addin projects have been added:
	* Gtk.Tasque.Columns: provides standard Tasque
columns
	* Gtk.Tasque.TimerCompleteColumns: provides
timed completion of tasks.

The functionality of those addins is the same as in the
previous version of Tasque. This change is about
making Gtk.Tasque more extensible and easier to
maintain.